[mlir][Arith] Extend unsupported float op emulation to `math` dialect
Because we may need platform-specific rewrites for `arith.extf` and
`arith.truncf` (see arith-to-amdgpu), extend the handling of
unsupported floating-ponit types to rewrite operations from the math
dialect, thus producing arith.extf/arith.truncf pairs.

This approach contrasts with the existing method of handling
low-presinios floats in the gpu-to-* rewrites.

I'm not entirely sure about the dependency I introduced here, but this
is an initial pass to get our usecase working.