[Attention] Use transposed/rotated attention kernel (#1354)
Before this commit, the attention kernel would look like as follows :
out = softmax(Q x K) x V

After this commit, it is now :
out = transpose( transpose(V) x softmax( transpose(K) x transpose(Q) ) )

This layout is interesting because for two reasons:

The reductions axis of the softmax will be laid out in just two threads for MFMA (hence 1344 is important)
It introduces an opportunity (not done here) to avoid LDS-based swizzling to prepare the operands to MFMA as the transposed thread layout is what the output is produced after the first gemm.
Additionally, this PR also includes relaxation of blocksize >= 512 check just for attention because the performant tuning config for triton problems are just that and happens to work fine. In future we can investigate, whether that check is relevant now.