Improve loads and stores involving scalar memrefs. (#1211)
Expand the scalar memref optimization the global load lowering had to
apply to threadwise_read_into operations from non-global memories.

Extend that optimization in global loads and stores to handle the case
where the tensor has 1 element, not just when it's 0D.

However, remove "assume it's in bounds" optimization from atomic
stores whose validity isn't trivially true. This is because, just
because all threads write to the same place doesn't mean they all have
valid data to atomically {add, max, ...} there.

We were probably going to trip over this eventually and waste a bunch
of time, hopefully this fix wards that off.

Improve the documentation on moveTransformsBefore() to make it more
obvious what that function does.

Adjust the fusion tiling flow to not create unrolled loops for what
are essentially fill/memset operations.