Fix incorrect logic around stores to length-1 memrefs
The code in #1211 was incorrect and led to colflicting stores to
memref<1x1x...x1xT>s that were padded out because it erroniously
ignored the validity bit.

This commit fixes the logic to always emit bounds checks in the
presence of non-trivial validity logic, even if the final coordinate
is 0D. This also improves the handling of such bounds checks by adding
expand_shape calls to create a coordinate that the buffer bounds check
trick can operate on.

(Honestly, it's surprising that this didn't come up with more
failures.)

Fixes https://github.com/ROCm/rocMLIR-internal/issues/1321