[Attention] Bypass LDS-based swizzling for second gemm's non-V input (#1388)
This PR conditionally enables the optimization that would avoid LDS-based swizzling for some cases of MFMA.
The detailed analysis is written here : ROCm/rocMLIR-internal#1201

In a summary, this works only for MFMA in KReduction mode (because there is no KReduction mode in WMMA) where we could match the operand thread layout with the MFMA output thread layout -- so that they could be used back-to-back without needing a LDS-based swizzling in the middle.

In order to keep the op verifiers calculate correct LDS requirement and also to enable rock/utility/ to be used everywhere where Rock IR is used (i.e. including rock/IR itself), I have moved all the utilities into rock/IR folder. This was required for gridwise_attention op verifier to calculate the correct LDS usage.