Fix transpose/collapse_shape soundness issues in TosaToRock
Fixes https://github.com/ROCm/AMDMIGraphX/issues/2692

Candidate for backport to 6.1.

The problematic transformation arises in the following context,
simplified from the original issue slightly:

%0 = expand_shape(%arg0, ...) : tensor<7x6x5xf32> to tensor<7x3x2x5xf32>
%1 = transpose(%0, [0, 2, 1, 3]) : tensor<7x3x2x5xf32> to %tensor<7x2x3x5xf32>
%2 = collapse_shape(%1, ...) : tensor<7x2x3x5xf32> to tensor<7x6x5xf32>

Previously, our code for rewriting transpose(collapse_shape) to simply
collapse_shape - which is important for ensuring we cleanly
pattern-match even when we're adding and removing unit dimensions all
over the place to match broadcasting requirments or operation
specifications - would rewrite such the code such that this entire
reshape/transpose/reshape chain was deleted.

That deletion led to incorrect results, because the within-dimension
tranposition that that sequence of operations specifies is
semantacially menaingful.

Due the complexity and pile of hacks that is
the collapse/trpnspose rewriter, I was unable to fully diagnose
exactly how we end up making this mistake in a reasonable time.

However, based on the reasonable assumption that we have this rewrite
to account for stray unit dimensions, I have simply added a guard such
that collapse_shape operations that perform non-trivial
concatentations (like the 2x3 -> 6 one above) cannot have transposes
folden into them, as we have no way to ensure that those dimensions
are not the results of some previous expand_shape or similar
operation.

This change required updating an existing test that relied on the old
behavior.