Fix collapseContiguousMerges() being too  greedy with unit dimensions (#1410)
Fixes https://github.com/ROCm/rocMLIR/issues/1407

The bug resolved here is one that comes from an overly aggressive approach
to adding length-1 dimensions to contiguous groups when detecting
Merge/Unmerge pairs, which led to unusual permutation sequences
like the one abstracted into the unit test for this commit
causing an overeager dimension collapse. This led to runtime page
faults because the H*W index into the output was being scaled by
the large multiplier for the batch (N) dimension of the result tensor,
leading to out of bounds accesses.

We guard against such issues in the future by strictly limiting the
unit dimension grabber to only include those unit dimensions that
lie within the contiguous group to be merged (as in the
original Merge{8, 1, 3}, Unmerge{8. 3} example).

In addition, we add rudimentary debug output for merge collapsing so that
it's clear that it's occurring.