[Attention] Implement a proper fusion between first gemm and softmax (#1399)
The current code for generating attention kernels only accepted two optional inputs
that could be used to mul (scale) and add (bias) between the first gemm and the softmax.

However, going forward we need squeeze in a arbitrary dequantization schemes that is
composed of element wise operations. Therefore, as a pre-requisite to that this PR
implements a generic fusion point between the first gemm and softmax.

The said tree is implemented as an inverted tree where the root of the tree is the softmax input
and one leaf being the first gemm output. Other leaves can reach farther as long as elementwise operations
go. Moreover, this PR add support to fused reshape-like ops as well because they are quite common
in graphs we get from MIGraphX that has rank > 3 and we do squeeze them to be rank = 3 around gemms because
tosa only support rank=3 matmuls.

Future work :

However, the code done here fundamentally can fuse any operator that lowers
to rock.transform, into the tree with 1 restriction : the path between the first gemm output leaf to the next branching point
in the tree should contain an invertible rock.transform.