Update device library finding to prepend ROCM_PATH, use hipconfig
Fixes #1415

Because our static library build wants to avoid runtime dependencies
on (the location of) ROCm, but we need the device libraries for math
functions, we pull the device libraries into a #include-able file
at configure time. A recent bug report showed the build system
picking up the wrong value for ROCM_PATH, leading to the device
libraries not being findable. This commit, hopefully, addresses
that issue.