Change vectorization and merge collapse APIs to take Value arguments
This commit removes getMaxVectorization[forDataType]() as a function
that takes a sequence of transform map attributes and replaces it with
a version that takes a transformed Value. Similarly,
collapseContiguousMerges() now takes the transformed Value and clones
the IR containing said contiguous merges in place instead of taking a
transform stack.

Furthermore, getMaxVectorization() now returns a VectorizationResult
struct, which includes the maximum vectorization fot the underlying
buffer along with information on how said underlying buffer is
naturally vectorized (this is currently unused, but will be needed for
rock.scalarize).

(While I was at it, I made the 128 bits / element type bitwidth thing
the default behavior for vectorization, since we almost always use it.)

This change is nedeed:
1. To enable getMaxVectorization() to reason about - and through - the
upcoming rock.scalarize operation, which isn't a coordinate
transformation but sort of acts like one.
2. To adapt the API for future dynamic shape cases, where a
TransformMapAttr might have symbols and so will need to, inherently,
be part of a TransformOp.

As a side effect, the vectorization inferenece tests now have nice
names and are split up into individual functions.