Fix incorrect indexing inte Merge parameters in -rock-fold-broadcast
Fixes #1417

FoldBroadcast would crash when processing Merge operations like
<Merge{4, 1} ["x"] at [1] -> ["a", "b"] at [1, 2]>, as it would index
the merge parameters with 1 and 2 and not 0 and 1, causing out of
bounds array accesses and thus assertion failures.

This commit resolves that issue.

In addition, while looking around nearby, I identifidied that the
handling for Broadcast{} operations that handled multiple
dimensions (like, say,
<Broadcast{1, 1} ["a", "b"] at [0, 1] -> ["a", "b"] at [0, 1]> )
was incorrect because the worklist update logic always checked the
last underlying dimension for Broadcast{}. This commit fixes that
issue.