Use `waves-per-eu` properly (#1439)
Currently waves-per-eu is calculated simply by blocksize / wavesize >= 2 heuristic.

I think this is not correct.
According to https://clang.llvm.org/docs/AttributeReference.html#amdgpu-waves-per-eu, we should only be doing this if there are sufficient parallelism within workgroup and/or grid constrained by the lds usage of the kernel. Otherwise, we will be losing scalar performance for non-existent occupancy.

This commit also removes bunch of side channelled info from higher up in the stack that were there to calculate lds usage where as in it could be properly calculated at rock-to-gpu before they are exported as symbols.