Filter out invalid WMMA perf configs (ex. m/nPerWave < 16)
If mPerWave or nPerWave are less than  the input length of the WMMA,
we can't cohenetly codegen usages of those WMMAs.

Similarly, while it wasn't an issue in this bug, if mPerWave or
nPerWave aren't evenly divisible by the wmma's input length, reject
that config as well.

Fixes https://github.com/ROCm/rocMLIR-internal/issues/1444

Move the tests