[Attention] Iterate on gemm1MBlocks (#1475)
Currently if gemm1MBlocks > 1 (which is rare) but is present in larger head_dim cases (i.e. 512),
we assign a single block in gemm1MBlocks to a seperate workgroup. This result in duplication of
first gemm and softmax for such splits.

This change instead of assigning a different workgroup. this will iterate on gemm1MBlocks in the same
workgroup.

This shows speed ups of 3x for f32 case and 67% for f16 case that is of interest to SDXL work.