[HOTFIX] Work around busted bf16 tests from ext/trunc folding (#1502)
The x86 backend now merges together sequences of the form
```llvm
%1 = fptrunc float %0 to bfloat
%2 = fpext bfloat %1 to float
...(%2)
```
to just `%0`.

Normally, this is fine, and just gives higher precision. However,
in our test code, when we're verifying a GPU reference (that is, one
where we did the computation in floats to test a lower-precision
datatype) we were relying on such trunc/ext pairs to lose the extra
precision (since the code under test returnd a memref<...xbfloat>, and
was extf'd to f32 for verification).

There is an arithmetic fence intrinsic that should let us prevent the
optimization in that one spot, but supporting it for bfloat on x86 is
a currently open PR. Until then, I've picked decently complex noop that
blocks this optimization.

Spceifically, for bfloat, we do
```mlir
%v = load %gpuRefOutput[...] : float
%t = arith.truncf %v : float to bfloat
%isneginf = arith.cmpf olt, %t, -BFLOAT_MAX
%t.clamp = arith.maximumf %t, -BFLOAT_MAX
%fix.neginf = select %isneginf, %t, %t.clamp
%vMasked = arith.extf %fix.neginf : bfloat to float
store %vMasked, %gpuRefOutput[...]
```

This'll work as a nice hotfix from what I can tell.