Fix excess devectorizaiton in buffer ops
The code that computed how many buffer ops were needed had the wrong
definition of "how many elements can go in one buffer load/store",
which led to, for example, loads of f16s only being vectorized to a
maximum of 2 elements.

The compiler conveniently patched this up for us in a lot of cases,
but while I was working on flat memrefs, I found instances where
the compiler didn't manage to be smart, causing performance regressions.

THis commit fixes this issue.