Refactor the host convolution harness to make it N-dimensional
Update the logic in the MLIR affine map based convolution test code to
make it work correctly on inputs of arbitrary length in preparation
for N-D convolution support in rocmlir-gen if we decide we want that.

(While we're here, make the upper bounds in the affine conditions the
constants that they are instead of symbols for no good reason)

Note that most of what would test this commit actually lives on the
flat memrefs branch since I had to restore it.