Using 1D arguments where posible (flat memrefs)
Initial testing revealed that LLVM produces noticeably (as in, measurable performance improvements) better code when the input and output buffers are given as 1D and we explicitly have an Unmerge{} to combine the N-dimensional logical tensor indices we're using.

The code changes that I expect are helping

We'd formerly do
```
%i0 = ...
%j0 = ...

...
%dI1 = ...
%dJ1 = ...
%i1 = add %i0, %dI1
%j1 = add %j0, %dJ1
%v = load %mem[%i1, %j1]
// Which is really
%tmp1_1 = mul %i1, J
%tmp2_1 = add %tmp1_1, %j1
%ptr_1 = gep %mem, %tmp2_1
%v = load %ptr_1
```

But now, with the explicit unmerge, we generate
```
%idx0 = %j0 + J * %i0
...
%dI1 = ...
%dJ1 = ...
%dI1_scaled = mul %dI1, J
%dIdx_1 = add %dI1_scaled, %dJ1
%idx1 = add %idx0, %didx_1
%v = load[%idx1]
```
which is a pattern that's much more legible to compilers (ex, you can hoist the gep for %idx0 outside the loop and the detector for that has a prayer of pattern matching)

This is especially notable in combination with

Currently, if we're trying to lower
```
%v = buffer_load %argN[%i0, %i1] if %c
```
we need to do (where L is the buffer extent)
```
%i0_if = select %c, %i0, 0
%i1_if = select %c, %i1, L
%idx = %i1_if + LEN(d1) * %i0_if
%desc = to_buffer_desc(%argN, L, ...)
amdgpu.buffer_load %desc, %idx, ...
```

Note those selects that are inputs to arithmetic. This isn't an obvious pattern, so it doesn't codegen well.

Now, when arguments are flat, we merely have
```
%idx_if = select %c, %idx, L
%desc = ...
```
which can much more obviously be implemented as a clamp or some other such simple operation.

Examination of an interesting gemm shows that this change does actually carry all the  way through to final codegen

1. Amends `rocmlir-gen` and `ConvGenerator.cpp` so that, when we generate a kernel for internal testing, the memref<>s passed in are 1-dimensional and then `rock.transform`'d to the shape that they used to have. Some care has been taken (except in attention) to ensure that the transforms have sensible dimension names
2. Changes the migraphx-to-tosa logic so that all tensor inputs and outputs that come from the MIGraphX dialect are 1-D and then `tosa.reshape`d back to their in-memory representation (which then goes to the existing logic for handling strides and slicing)
3. Removes the old C++ convolution verifier which is unused and would've been a pain to generalize.

Built-in tests have been amended to expect 1-D memref printing when the underlying code was changed to 1D memrefs.

Existing tosa-level tests have not been 1-D-ified.

Existing Rock-level tests still work on N-D memrefs in most cases

Tests for the structure of the CPU convolution and for testing that
the verifier produces expected output have been **restored** and
strengthened.

The C++ verifier was our old infrastructure for verifying convolutions before we moved to generating the affine loop nests in rocmlir-gen. The verifier only worked for 2D convolutions.

It has been meaningfully dead (only tested by its own tests) for over a year. Refactoring that code to support the new calling convention (flat memrefs) would be a lot of work for zero gain, and thus, this commit goes and deletes the darn thing.