Manually set overflow flags when expanding indexing maps
When we work with transform maps in the context of rocMLIR, we know
the following properties that are not true of a general affine map:

1. We may assume that the inputs to the map are non-negative. If a
previous computation (like subtracting off padding) would cause the
inputs to potentially be negative, the validity bit will cause us to
not use the final computation result, so we can have all the undefined
behavior/unjustified optimizations/... in the rest of the address
calculation - it won't matter.

To put it another way, we speculatively execute all the index
calculatinos for the valid case, and then discard the result in the
invalid case.

LLVM can't know that's what we meant, so we have to do that annocation
ourselves.

2. During the index diffs mechanism, the individual diffs can be
negative (ex. if our map is "x -> x / 3, x % 3" and we have the inital
coordinates (0, 2), then moving to x + 1 gets us coordinates (1, 0)
and diffs (1, -2)). We also can therefore see a diff becoming negative
in general. However, we know that, unless we're in an operator that
can impact validity (Pad{}, mainly, but also some Embed{}), the sum of
the original index and the diff is always non-negative.

That non-negative-ness allows us to sprinkle `nuw` on most index diff
updates, which is also a property that might be hard to infer in
general.

--

This entire setup is implemented with a custom affine map expanded,
which has the side benefit of letting us skip emitting all that super
noise code for `mod`. That visitor looks for potentially negative
expressions (that is, those with negative constants somewhere in their
subtrees) and annotates all the adds and multiplies with nuw/nsw as
appropriate for the non-negativity of the inputs.

(Note, because of the analysis up in AnalyzeMemoryUse, we have the
invariant that signed overflow can't occur - if you ever need x
int32_t_max as a map output, we go to 64-bit `index` instead).