Implemented support for the update.dpp intrinsic, replacing the previous mov.dpp usage. (#1467)
* "Extended the DPP operation to support the optional 'old' operand
and the update.dpp intrinsic. Implemented and ensured 'old'
defaults to 'src' when not provided. This allows the use of
update.dpp in all cases, effectively behaving like mov.dpp when
lowering in backend."

* "Updated AMDGPU_DPPOp to make 'old' operand mandatory and replaced
usage of mov.dpp with update.dpp due to deprecation of mov.dpp.
Added documentation in the IntrinsicsAMDGPU.td file to reflect
these changes and guide future usage."

* Resolved deprecation warnings by updating usage of dyn_cast in attribute retrieval.

* "Added support for 64-bit operands for update dpp operation."