Restore the lit.local.cfg files for host-code tests
The host code tests can't run in cases where we don't have a JIT
available, or, equivalently, in the fat library build. I managed to
rescue the tests themselves from old branches, but didn't bring the
guards with them, causing the static library build to fail. This
commit fixes that.