Fix weekly Jenkins stalls by removing "any" machine selection (#1541)
According to @okakarpa, the "any" machine would translate to a jenkins server that doesn't have any label. Since @okakarpa has made sure every machine will have a label, this will stall the weekly test pipeline forever. The workaround is to add a `build-only` label so the result can be archived and pipeline proceed.