[MHAL] Update MHAL to use gpu target attrs: SerializeToHSA to TargetAttr (2/3)
This patch updates the MHAL project to use the `gpu` target attributes
infrastructure. This is a patch in part of the series switching the compilation
infrastructure from SerializeToHSA to `gpu` target attributes.

This includes:
- Updating `PrefillPass` to work on GPU binaries.
- Updating `PackageTargetsPass` to work on GPU binaries.
- Updating `MHALToGPU` to create `gpu.binary` operations.
- Updating `mhal::TargetObject` to store an attribute instead of a string.
- Add the `DropMetadata` pass:
This pass drops all metadata from GPU binaries -e.g. property dictionary and
kernel metadata. This is required to avoid unregistered dialect errors
for attributes stored in the metadata like `mhal.prefill` but not parsable
by tools like `mlir-cpu-runner`.