[roc] Update roc to use gpu target attrs: SerializeToHSA to TargetAttr (3/3)
This patch updates the `rocMLIR` project to use the `gpu` target attributes
infrastructure. This is a patch in part of the series switching the compilation
infrastructure from SerializeToHSA to `gpu` target attributes.

This includes:
- Removing unnecessary headers and deprecated serialization passes.
- Adding new compilation passes to the rocMLIR kernel pipeline.
- Updating `CheckResidencyPass` to use `gpu.binary` ops.
This change also requires the pass to be run at the module level, as
`gpu.binary` ops are not isolated from above.
- Updating CAPI `mlirGetBinary` function to work on gpu binaries.
- Updating `rocmlir-lib` to work on `gpu.binary` ops.
- Adding missing gpu compilation registration calls to `xmir-runner`.
- Updating tests to work with the target attribute compilation mechanism.
- Updating the tunning driver to work on gpu binaries.
- Adding a target interface for the `ROCDLTargetAttr`:

This change introduces a custom `gpu::TargetAttrInterface` for the
`ROCDLTargetAttr` attribute. This interface overrides the default
behavior of:
`mlir::ROCDL::registerROCDLTargetInterfaceExternalModels`
By adding:
 - rocMLIR specific metadata in the binary.
 - Allowing to package device bitcode libs inside the rocMLI binaries.
 - Use LLD API instead of calling the executable.

To register this interface one must call:
`mlir::registerRocTarget`.

The idea of this custom interface is to minimize collisions with upstream.