Rearrange alloc_tensor creation and CSE to fix input fusion crash
Fixes https://github.com/ROCm/rocMLIR-internal/issues/1520

As mentioned in the comments in the code below, this PR fixes an issue
that areses because `tensor.empty` operations were combined in a way
that caused the regularization pass to trip an assert.

To put it another way, the rocMLIR fusion code relies on never needing
to try to "fuse"

%a = alloc()
f(reads ..., writes %a)
%b = alloc()
g(reads ... + %a, writes %b)
// No other uses of %a until
h(reads ... + %b, writes %a)

This can happen if you combine two tensor.empty ops and the
buferization analysis notices that you don't actually need two buffers
for the two different empty tensors.

Therefore, to fix the assert, as a workaround, copy the translation
from tensor.empty to bufferization.alloc_tensor before the call to
CSE. We leave the existing call to that pass in place just to be safe.

This has a side effect of increasing the number of generics in the
tosa transpsoe optimiization test.