[CI] Make our PR CI use a fixed layout for conv
The reason we have a tuning of a single conv in PR CI
to make sure the tuner is functional.

To that effect, I think we dont have to tune for all layouts
because we have that coverage in e2e tests.

This commit is to fix the layout that we tune in PR CI
so it would be not painful on slower HW (Navi3x).