Add reduce as a fusor in regularization
rock.reduce is an operation that can write intermediate buffers and
get fused in to things, just like linalg.generic. However, it wasn't
being treated as a fusor by the regularization pass, breaking the "no
transforms on readers" invariant that the gridwise-gemm-to-blockwise
..= align-tiling pipeline relies on.

This patch fixes this issue.

Thanks to @manupak in PR #1579 for the initial approach and the tests

Fixes https://github.com/ROCm/rocMLIR-internal/issues/1541