[mhal] Add -mhal-emulate-narrow-type pass
Add a pass that coordinates upstream's rewrite patterns for emulating
memrefs of narrow types (primarily i4) and adds patterns that allow
rewriting functions that take i4 memrefs as arguments.

The pattern-creation functions are exposed to enable reusing them in
the Rock analogue of this pass.

This pass only rewrites scalar memref operations because that's all
that's needed for complinig host/test code.c