Add rewrites for migraphx.unpack to the rocMLIR pipeline
This commit handles MIGraphX's unpack operation (which is used to
unfold tensors of packed 4-bit integers into 8-bit integer tensors) by
exposing the underlying int4 types. This means that future operatons
don't have to deal with an expanding fusion, that the type system
knows the real type of all tensors, and that the upstream int4
load/store emulation works correctly.

To accomplish this, the unpack operations gets rewriten into on that
unpacks from a byte buffer into a tensor of int4 (adjusting the shape
nad stride appropriately) which is then turned into the underlying
byte buffer with a zero-extension (a newly added attribute) and a
reshape.

Then, patterns fold away unpack, turning transpose(unpack(x)) into
unpack(transpose(x)) (and similarly for reshape) until the unpack()
reaches the function border. There, the byte-tensor argument is
replaced with one that uses an int4 tensor.

Note that this is the first commit, I haven't written all the tests
yet, but I want to unblock MIGraphX.