Handle non-zero-preserving input fusions, make read_into track validity
Apologies for the mega-commit.

Previously, if you input-fused an elementwise funciton like
(x) => (x + 4), this would cause issues, because the 4 would be added
both to elements of the input value but also to padding values
introduces during tiling, which would cause incorrect results.

To fix this, we:
1. Make threadwise_read_into optionally return a boolean vector of
validities, recording whether or not an element in the register tile
actually came from memory or not
2. Give threadwise_read_into a dynamic validities argument, where the
validity data from the coordinate transforms is ANDed with the
validity vectors from those arguments to allow applying validitiies
that don't come from coordinate transforms.
3. Update all the threadwise_read_into operations that can have input
funions on them to generate a validity record.
4. It a linalg.generic is being tiled by a threadwise_read_into and
that read records validity (that is, it came from an input
fusion) *and* the result of the elementwise function on all-zero
inputs isn't zero, we propagate the validities.

That is, we make the threadwise_read_into ops that read each input to
the generic record their validities, and then do a register ->
register threadwise_read_into (which would ordinarily be a memcpy())
with dynamic validities taken from those reads to re-apply the 0 mask
after the generic runs.