Move threadwise_copy ops in gridwise_gemm_accel, pipeline non-accel (#1603)
1. The threadwise_copy ops were mistakenly placed near the
threadwise_read_into ops when the pipelined loop was set up for
gridwise_gemm_accel. Fix this by moving them into the right spot in
the pipelined loop so that we don't have `shufflevector`s forcing
global loads to complete

2. Change the non-accel lowering path tho use the software pipelining
scheme so that it matches the accel path, making future fusion
optimizations easier.