Fix crash arising from insufficient guards in WMMA instruction selector (#1621)
When serecting WMMA instructions, we get the number of repeats by
dividing by dPerAccel. We used to guard this against division by 0 by
checking that mPerWave and nPerWave were divisible by inputVectorLen,
which was equal to dPerAccel ... until gfx12, where dPerAccel becase
double the length of inputVectorLen (because WMMA changed to re-use
both halves of the wave). This caused 0 coefficients in transform
maps, which led to a bunch of weird crashes.

Fixes https://github.com/ROCm/rocMLIR-internal/issues/1582 (by making
the input problem invalid) but doesn't actually fix the assertion
crash we saw in that ticket because I got a different crash on
develop.