MIGraphX changes for int4 (#1596)
* Define migraphx.unpack operator, add unsigned integer conversions

We define migraphx.unpack to mirror the operaton MIGraphX uses to
decompress int4 buffers.

We add a `zeroExtend` option to `migraphx.convert` that doesn't exist
in MIGraphX so that we can represent the unpacking in a way that makes
more sense to us.

* Add rewrites for migraphx.unpack to the rocMLIR pipeline

This commit handles MIGraphX's unpack operation (which is used to
unfold tensors of packed 4-bit integers into 8-bit integer tensors) by
exposing the underlying int4 types. This means that future operatons
don't have to deal with an expanding fusion, that the type system
knows the real type of all tensors, and that the upstream int4
load/store emulation works correctly.

To accomplish this, the unpack operations gets rewriten into on that
unpacks from a byte buffer into a tensor of int4 (adjusting the shape
nad stride appropriately) which is then turned into the underlying
byte buffer with a zero-extension (a newly added attribute) and a
reshape.

Then, patterns fold away unpack, turning transpose(unpack(x)) into
unpack(transpose(x)) (and similarly for reshape) until the unpack()
reaches the function border. There, the byte-tensor argument is
replaced with one that uses an int4 tensor.

Note that this is the first commit, I haven't written all the tests
yet, but I want to unblock MIGraphX.

* Move to custom op, no tests, no rewrite

* Pass to lower custom Toas ops to linalg

* Add tests, correct minor bugs revealed by said tests

* Simplify the lowering of int4

Given that we've got model-level guarantees that there'll be no
nonsense where the bytes are transposed - or at least, if they are,
there'll likely be explicit reshapes about it - we can just use a
ddouble-length int4 axis

* int4 e2e test, might not work, machine acting up

* Fix e2e tests, add the right documentation

* Rename pass to rocmlir-custom-tosa-to-linalg

* Add an isUnsigned flag to migraphx.unpack

* Multibroadcast

* Update e2e test I didn't notice the issue with