Fix overly-strict guards in LLVM conversions for fp8 intrinsic
After fp8 conversions were lowered to AMDGPU dialect ops, those
operations were not being converted to the LLVM intrinsics they
correspond to because the rewrite patterns were still checking for
gfx940+.

As part of this, factor out tests for type-match isto isNativeFp8()
and isNativeBf8() functions in the AMDGPUToRocdl rewrites.

Also, fix a typo in isGfx940() that caused it to be true for gfx950.

Finally, test all these OCP format conversions by duplicating the
gfx940 tests.