[mlir][AMDGPU] Remove an old bf16 workaround
The AMDGPU backend now implements LLVM's `bfloat` type. Therefore, we
no longer need to type convert MLIR's `bf16` to `i16` during lowerings
to ROCDL.

As a result of this change, we discovered that, whel the code for MFMA
and WMMA intrinsics was mainly prepared for this change, we were
failing to bitcast the bf16 results of WMMA operations out from the
i16 they're natively represented as. This commit also fixes that issue.