[MHAL] Move read_access and write_access to MHAL
Move the annotating read_access and write_access to mhal. instead of
func. so we're not carrying unneeded upstream patches.

This means moving the arg attributes into mhal and removing mention of
them from the tosa partitioner/outliner. Therefore, we need to add a
pass to MHAL that adds the read_access or write_access to tensor
arguments the way the outliner used to do.