Add support for `pad` + removeSubDims in `removeUpperDims` (#1639)
Currently, the pad is not properly handled when removing upper dims,
especially in presense of dimensions collapse and expansion happenning.

This PR adds support for that in terms of the two cases that be there:

1.) If a part of the dimension that is being removed does not involve a padded region :
this means we can do the padding as usual as the padding is contained within the subtile.

2.) If a part of the dimension that is being removed does involve a padded region :
From the subtile's perspective, this means we are taking a tile out of a dimension where edges are padded. Therefore, as long as the subtile is considered the padding does not affect coordinates within the subtile because they would be materialized with the padded (zero most of the time) value. Thus, we ll ignore the padding.