Fix test include path and replace LLVM internal files with HSA
Some arch values are not available in HIP, se we tried to fetch
them from LLVM internal files (GCNSubtarget). This however casued
build (and CI) failures.

Instead we can rely on HSA to get most of the information, thus
avoiding LLVM internal files. In other words, this now should
build succesfully.

There are still some fields (commented and with ???) that cannot
be retrieved by either HIP or HSA - we will need to investigate
what to do with those.