Check for multiple fusion root ops  (#2042)
It is possible that sometimes TosaToRock or MIGraphX generates IR such that it produces invalid IR where a single function has multiple FusionRoot ops. Root ops are either GemmLike or GemmPlusGemmLike. This is not properly supported in rocMLIR right now. It may run correctly for some cases though. But untill we officially declare and maintain support for such cases, it is better to detect such cases and throw error.